Refactored to move the SharingModel methods to InputFileExtensions and renamed to make the connection to SharingModel more explicit. Refactored the Deserialize method to use the GetMemoryStream method available in InputFileExtensions.
